Split the acceptance tests
