+++ src/it/scala/org/zalando/spearheads/innkeeper/AcceptanceSpecsHelper.scala
@@ -1,0 +1,71 @@
+package org.zalando.spearheads.innkeeper
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model._
+import akka.http.scaladsl.model.headers.{OAuth2BearerToken, Authorization}
+import akka.stream.ActorMaterializer
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.time.{Millis, Seconds, Span}
+import scala.collection.immutable.Seq
+
+/**
+  * @author dpersa
+  */
+object AcceptanceSpecsHelper extends ScalaFutures {
+
+  val uri = "http://localhost:8080/routes"
+  implicit val defaultPatience =
+    PatienceConfig(timeout = Span(2, Seconds), interval = Span(500, Millis))
+  implicit val system = ActorSystem("main-actor-system")
+  implicit val materializer = ActorMaterializer()
+
+  def entityString(response: HttpResponse): String = {
+    response.entity.dataBytes
+      .map(bs => bs.utf8String)
+      .runFold("")((a, b) => a + b)
+      .futureValue
+  }
+
+  def postSlashRoutes(routeType: String)(token: String): HttpResponse = {
+    val route =
+      s"""{
+          |  "name": "THE_ROUTE",
+          |  "description": "this is a route",
+          |  "activate_at": "2015-10-10T10:10:10",
+          |  "route": {
+          |    "matcher": {
+          |      "path_matcher": {
+          |        "match": "/hello-*",
+          |        "type": "${routeType}"
+          |      }
+          |    }
+          |  }
+          |}""".stripMargin
+
+    val entity = HttpEntity(ContentType(MediaTypes.`application/json`), route)
+
+    val headers = Seq[HttpHeader](Authorization(OAuth2BearerToken(token)))
+
+    val request = HttpRequest(method = HttpMethods.POST,
+      uri = uri,
+      entity = entity,
+      headers = headers)
+
+    val futureResponse = Http().singleRequest(request)
+    futureResponse.futureValue
+  }
+
+  def getSlashRoutes(token: String): HttpResponse = {
+    val futureResponse = Http().singleRequest(HttpRequest(uri = uri,
+      headers = Seq[HttpHeader](Authorization(OAuth2BearerToken(token)))))
+    futureResponse.futureValue
+  }
+}
+
+object AcceptanceSpecTokens {
+  val READ_TOKEN = "token-employees-route.read"
+  val WRITE_STRICT_TOKEN = "token-employees-route.write_strict"
+  val WRITE_REGEX_TOKEN = "token-employees-route.write_regex"
+  val INVALID_TOKEN = "invalid"
+}
